Move factory to magnet.factory
