+++ siriuspy/tests/magnet/test_factory.py
@@ -4,13 +4,13 @@
 
 from unittest import mock, TestCase
 import siriuspy.util as util
-import siriuspy.factory as factory
-from siriuspy.factory import NormalizerFactory
+import siriuspy.magnet.factory as factory
+from siriuspy.magnet.factory import NormalizerFactory
 from siriuspy.magnet.normalizer \
     import DipoleNormalizer, MagnetNormalizer, TrimNormalizer
 
 
-valid_interface = (
+PUB_INTERFACE = (
     'NormalizerFactory',
 )
 
@@ -21,10 +21,10 @@ class TestMagnetFactory(TestCase):
 
     def test_public_interface(self):
         """Test module's public interface."""
-        valid = util.check_public_interface_namespace(factory, valid_interface)
+        valid = util.check_public_interface_namespace(factory, PUB_INTERFACE)
         self.assertTrue(valid)
 
-    @mock.patch('siriuspy.factory._norm.DipoleNormalizer', autospec=True)
+    @mock.patch('siriuspy.magnet.factory._norm.DipoleNormalizer', autospec=True)
     def test_dipole_creation(self, mock_data):
         """Test Factory.create. dipole creation."""
         maname = 'SI-Fam:MA-B1B2'
@@ -32,7 +32,7 @@ class TestMagnetFactory(TestCase):
         magnet = NormalizerFactory.create(maname=maname)
         self.assertIsInstance(magnet, DipoleNormalizer)
 
-    @mock.patch('siriuspy.factory._norm.MagnetNormalizer', autospec=True)
+    @mock.patch('siriuspy.magnet.factory._norm.MagnetNormalizer', autospec=True)
     def test_magnet_creation(self, mock_data):
         """Test Factory.create magnet creation."""
         maname = 'SI-Fam:MA-QDA'
@@ -40,7 +40,7 @@ class TestMagnetFactory(TestCase):
         magnet = NormalizerFactory.create(maname=maname)
         self.assertIsInstance(magnet, MagnetNormalizer)
 
-    @mock.patch('siriuspy.factory._norm.MagnetNormalizer', autospec=True)
+    @mock.patch('siriuspy.magnet.factory._norm.MagnetNormalizer', autospec=True)
     def test_pulsed_magnet_creation(self, mock_data):
         """Test Factory.create magnet creation."""
         maname = 'SI-01SA:PM-InjDpKckr'
@@ -48,7 +48,7 @@ class TestMagnetFactory(TestCase):
         magnet = NormalizerFactory.create(maname=maname)
         self.assertIsInstance(magnet, MagnetNormalizer)
 
-    @mock.patch('siriuspy.factory._norm.TrimNormalizer', autospec=True)
+    @mock.patch('siriuspy.magnet.factory._norm.TrimNormalizer', autospec=True)
     def test_trim_creation(self, mock_data):
         """Test Factory.create trim creation."""
         maname = 'SI-01M1:MA-QDA'
